This project is part of Business Analytics Nanodegree at Udacity learning platform.
In this project, you'll create visualizations to reveal insights from a data set. You will create data visualizations that tell a story or highlight patterns in the data set. Your work should be a reflection of the theory and practice of data visualization, such as visual encodings, design principles, and effective communication. Tableau will be used to conduct the dashboards. 

There are 3 different data sets you can choose from.

    Flight Delays and Cancellations
    US Census Demographic Data
    Youtube Data from the US